+++ src/main/java/com/github/t45k/abcd/ast/constructor/AbstractASTConstructor.java
@@ -1,4 +1,4 @@
-package com.github.t45k.abcd.ast;
+package com.github.t45k.abcd.ast.constructor;
 
 import java.nio.file.Path;
