+++ src/main/java/com/github/t45k/abcd/ast/constructor/BindingASTConstructor.java
@@ -1,4 +1,4 @@
-package com.github.t45k.abcd.ast;
+package com.github.t45k.abcd.ast.constructor;
 
 import java.nio.file.Path;
